moved consts.py to terraform package added RESOLVED_MODULE_ENTRY_NAME const
