--- checkov/common/util/consts.py
@@ -1,1 +1,0 @@
-DEFAULT_EXTERNAL_MODULES_DIR = '.external_modules'
