+++ tests/terraform/parser/test_parser_modules.py
@@ -2,7 +2,7 @@ import os
 import shutil
 import unittest
 
-from checkov.common.util.consts import DEFAULT_EXTERNAL_MODULES_DIR
+from checkov.terraform.consts import DEFAULT_EXTERNAL_MODULES_DIR
 from checkov.terraform.parser import Parser
